requirements.txt

fastapi
uvicorn
openai-whisper
moviepy
sumy
python-multipart
